Refactors and restructures frontend code
Improves code organization and maintainability by renaming and relocating components.

Updates import paths and project structure for clarity.

Addresses ESLint errors and warnings to ensure code quality.

Introduces a new ESLint configuration file and adjusts relevant dependencies.
