+++ frontend/src/app/App.jsx
@@ -1,23 +1,23 @@
-import "./App.css"
-import { Routes, Route } from "react-router-dom"
-import LandingPage from "../pages/public/LandingPage"
-import HomePage from "../pages/dashboard/HomePage"
-import LoginPage from "../pages/auth/LoginPage"
-import RegisterPage from "../pages/auth/RegisterPage"
-import ForgotPasswordPage from "../pages/auth/ForgotPasswordPage"
-import OTPVerificationPage from "../pages/auth/OTPVerificationPage"
-import ChangePasswordPage from "../pages/auth/ChangePasswordPage"
-import AboutPage from "../pages/public/AboutPage/index.jsx"
+import { Route, Routes } from "react-router-dom"
 import ProtectedRoute from "../guards/ProtectedRoute.jsx"
-import MarketplacePage from "../pages/Marketplace/MarketplacePage"
-import PostDetailPage from "../pages/Marketplace/PostDetailPage"
+import MyPostsPage from "../pages/Dashboard/MyPostsPage"
+import HomePage from "../pages/HomePage/index.jsx"
 import CreatePostPage from "../pages/Marketplace/CreatePost"
 import EditPostPage from "../pages/Marketplace/EditPostPage"
-import MyPostsPage from "../pages/Dashboard/MyPostsPage"
+import MarketplacePage from "../pages/Marketplace/MarketplacePage"
+import PostDetailPage from "../pages/Marketplace/PostDetailPage"
+import ChangePasswordPage from "../pages/auth/ChangePasswordPage"
+import ForgotPasswordPage from "../pages/auth/ForgotPasswordPage"
+import LoginPage from "../pages/auth/LoginPage/index.jsx"
+import OTPVerificationPage from "../pages/auth/OTPVerificationPage"
+import RegisterPage from "../pages/auth/RegisterPage"
 import ArtistProfile from "../pages/profiles/ArtistProfile.jsx"
 import ClientProfile from "../pages/profiles/ClientProfile.jsx"
 import EditProfile from "../pages/profiles/EditProfile.jsx"
 import ProfileRouter from "../pages/profiles/ProfileRouter.jsx"
+import AboutPage from "../pages/public/AboutPage/index.jsx"
+import LandingPage from "../pages/public/LandingPage/index.jsx"
+import "./App.css"
 
 function App() {
   return (
