+++ frontend/src/components/layout/AuthFooter.jsx
