+++ frontend/src/components/layout/PublicFooter.jsx
