+++ frontend/src/components/layout/PublicNavbar.jsx
