+++ frontend/src/components/layout/SharedPublicFooter.jsx
