--- frontend/src/components/marketplace/CreatePostFormNew.jsx
@@ -1,437 +1,0 @@
-"use client"
-
-import React, { useState } from "react"
-import { useNavigate } from "react-router-dom"
-import { useAppDispatch } from "../../hook/useRedux"
-import { createPost } from "../../store/slices/marketplaceSlice"
-import { Trash2, Upload, X } from "lucide-react"
-import Loader from "../ui/Loader"
-
-const CreatePostForm = ({ onSubmit, onCancel, isSubmitting = false }) => {
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-    budgetMin: "",
-    budgetMax: "",
-    duration: "",
-    location: "",
-    skills: [],
-    category: "",
-    photos: [],
-  })
-
-  const [newSkill, setNewSkill] = useState("")
-  const [uploadingImages, setUploadingImages] = useState(false)
-
-  const categories = [
-    "Graphic Design",
-    "Web Design",
-    "Logo Design",
-    "Illustration",
-    "Photography",
-    "Video Editing",
-    "Animation",
-    "UI/UX Design",
-  ]
-
-  const popularSkills = [
-    "Photoshop",
-    "Illustrator",
-    "Figma",
-    "After Effects",
-    "Premiere Pro",
-    "Sketch",
-    "InDesign",
-    "Blender",
-    "HTML/CSS",
-    "JavaScript",
-    "React",
-    "Vue.js",
-  ]
-
-  const handleInputChange = (e) => {
-    const { name, value } = e.target
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }))
-  }
-
-  const handleSkillToggle = (skill) => {
-    setFormData((prev) => ({
-      ...prev,
-      skills: prev.skills.includes(skill)
-        ? prev.skills.filter((s) => s !== skill)
-        : [...prev.skills, skill],
-    }))
-  }
-
-  const handleAddCustomSkill = () => {
-    if (newSkill.trim() && !formData.skills.includes(newSkill.trim())) {
-      setFormData((prev) => ({
-        ...prev,
-        skills: [...prev.skills, newSkill.trim()],
-      }))
-      setNewSkill("")
-    }
-  }
-
-  const handleRemoveSkill = (skillToRemove) => {
-    setFormData((prev) => ({
-      ...prev,
-      skills: prev.skills.filter((skill) => skill !== skillToRemove),
-    }))
-  }
-
-  const handleImageUpload = async (e) => {
-    const files = Array.from(e.target.files)
-    if (files.length === 0) return
-
-    setUploadingImages(true)
-    try {
-      const uploadedImages = await uploadImages(files)
-      setFormData((prev) => ({
-        ...prev,
-        photos: [...prev.photos, ...uploadedImages],
-      }))
-    } catch (error) {
-      console.error("Image upload failed:", error)
-      alert("Failed to upload images. Please try again.")
-    } finally {
-      setUploadingImages(false)
-    }
-  }
-
-  const handleRemoveImage = (index) => {
-    setFormData((prev) => ({
-      ...prev,
-      photos: prev.photos.filter((_, i) => i !== index),
-    }))
-  }
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-
-    // Validation
-    if (!formData.title || !formData.description || !formData.category) {
-      alert("Please fill in all required fields")
-      return
-    }
-
-    if (formData.skills.length === 0) {
-      alert("Please add at least one skill")
-      return
-    }
-
-    // Prepare data for submission
-    const postData = {
-      ...formData,
-      budgetMin: parseFloat(formData.budgetMin) || 0,
-      budgetMax: parseFloat(formData.budgetMax) || 0,
-    }
-
-    onSubmit(postData)
-  }
-
-  return (
-    <div className="bg-gray-800/20 backdrop-blur border border-gray-700/50 rounded-xl p-8">
-      <form onSubmit={handleSubmit} className="space-y-8">
-        {/* Basic Information */}
-        <div className="space-y-6">
-          <h3 className="text-xl font-semibold text-white mb-4 flex items-center">
-            <span className="w-2 h-2 bg-[#A95BAB] rounded-full mr-3"></span>
-            Basic Information
-          </h3>
-
-          <div className="space-y-2">
-            <label htmlFor="title" className="block text-sm font-medium text-gray-300">
-              Project Title *
-            </label>
-            <input
-              type="text"
-              id="title"
-              name="title"
-              value={formData.title}
-              onChange={handleInputChange}
-              placeholder="e.g., Professional Logo Design Available"
-              className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-              required
-            />
-          </div>
-
-          <div className="space-y-2">
-            <label htmlFor="category" className="block text-sm font-medium text-gray-300">
-              Category *
-            </label>
-            <select 
-              id="category" 
-              name="category" 
-              value={formData.category} 
-              onChange={handleInputChange} 
-              className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-              required
-            >
-              <option value="">Select a category</option>
-              {categories.map((category) => (
-                <option key={category} value={category}>
-                  {category}
-                </option>
-              ))}
-            </select>
-          </div>
-
-          <div className="space-y-2">
-            <label htmlFor="description" className="block text-sm font-medium text-gray-300">
-              Project Description *
-            </label>
-            <textarea
-              id="description"
-              name="description"
-              value={formData.description}
-              onChange={handleInputChange}
-              placeholder="Describe your services, experience, and what makes you unique..."
-              rows={6}
-              className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all resize-none"
-              required
-            />
-          </div>
-        </div>
-
-        {/* Pricing & Timeline */}
-        <div className="space-y-6">
-          <h3 className="text-xl font-semibold text-white mb-4 flex items-center">
-            <span className="w-2 h-2 bg-[#A95BAB] rounded-full mr-3"></span>
-            Pricing & Timeline
-          </h3>
-
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-            <div className="space-y-2">
-              <label htmlFor="budgetMin" className="block text-sm font-medium text-gray-300">
-                <DollarSign className="inline w-4 h-4 mr-1" />
-                Minimum Budget
-              </label>
-              <input
-                type="number"
-                id="budgetMin"
-                name="budgetMin"
-                value={formData.budgetMin}
-                onChange={handleInputChange}
-                placeholder="0"
-                min="0"
-                step="50"
-                className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-              />
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="budgetMax" className="block text-sm font-medium text-gray-300">
-                <DollarSign className="inline w-4 h-4 mr-1" />
-                Maximum Budget
-              </label>
-              <input
-                type="number"
-                id="budgetMax"
-                name="budgetMax"
-                value={formData.budgetMax}
-                onChange={handleInputChange}
-                placeholder="0"
-                min="0"
-                step="50"
-                className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-              />
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <label htmlFor="duration" className="block text-sm font-medium text-gray-300">
-              <Clock className="inline w-4 h-4 mr-1" />
-              Project Duration
-            </label>
-            <input
-              type="text"
-              id="duration"
-              name="duration"
-              value={formData.duration}
-              onChange={handleInputChange}
-              placeholder="e.g., 1 week, 2-3 days, 1 month"
-              className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-            />
-          </div>
-
-          <div className="space-y-2">
-            <label htmlFor="location" className="block text-sm font-medium text-gray-300">
-              <MapPin className="inline w-4 h-4 mr-1" />
-              Location
-            </label>
-            <input
-              type="text"
-              id="location"
-              name="location"
-              value={formData.location}
-              onChange={handleInputChange}
-              placeholder="e.g., Remote, New York, Worldwide"
-              className="w-full px-4 py-3 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-            />
-          </div>
-        </div>
-
-        {/* Skills */}
-        <div className="space-y-6">
-          <h3 className="text-xl font-semibold text-white mb-4 flex items-center">
-            <span className="w-2 h-2 bg-[#A95BAB] rounded-full mr-3"></span>
-            Skills & Expertise
-          </h3>
-
-          <div className="space-y-4">
-            <div>
-              <p className="text-sm font-medium text-gray-300 mb-3">Popular Skills</p>
-              <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-2">
-                {popularSkills.map((skill) => (
-                  <button
-                    key={skill}
-                    type="button"
-                    onClick={() => handleSkillToggle(skill)}
-                    className={`px-3 py-2 rounded-lg text-sm font-medium transition-all duration-300 transform hover:scale-105 ${
-                      formData.skills.includes(skill)
-                        ? "bg-[#A95BAB] text-white shadow-lg shadow-[#A95BAB]/25"
-                        : "bg-gray-700/50 text-gray-300 hover:bg-gray-600/50 hover:text-white"
-                    }`}
-                  >
-                    {skill}
-                  </button>
-                ))}
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <p className="text-sm font-medium text-gray-300">Add Custom Skill</p>
-              <div className="flex gap-2">
-                <input
-                  type="text"
-                  value={newSkill}
-                  onChange={(e) => setNewSkill(e.target.value)}
-                  placeholder="Enter a skill..."
-                  className="flex-1 px-4 py-2 bg-gray-700/50 border border-gray-600/50 rounded-lg text-white placeholder-gray-400 focus:ring-2 focus:ring-[#A95BAB] focus:border-[#A95BAB] transition-all"
-                />
-                <button
-                  type="button"
-                  onClick={handleAddCustomSkill}
-                  className="px-4 py-2 bg-[#A95BAB] hover:bg-[#A95BAB]/80 text-white rounded-lg transition-all duration-300 transform hover:scale-105"
-                >
-                  <Plus className="w-4 h-4" />
-                </button>
-              </div>
-            </div>
-
-            {formData.skills.length > 0 && (
-              <div className="space-y-2">
-                <p className="text-sm font-medium text-gray-300">Selected Skills</p>
-                <div className="flex flex-wrap gap-2">
-                  {formData.skills.map((skill) => (
-                    <span
-                      key={skill}
-                      className="inline-flex items-center space-x-2 px-3 py-1 bg-[#A95BAB] text-white rounded-full text-sm font-medium"
-                    >
-                      <span>{skill}</span>
-                      <button
-                        type="button"
-                        onClick={() => handleRemoveSkill(skill)}
-                        className="hover:bg-white/20 rounded-full p-0.5 transition-colors"
-                      >
-                        <X className="w-3 h-3" />
-                      </button>
-                    </span>
-                  ))}
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Portfolio Images */}
-        <div className="space-y-6">
-          <h3 className="text-xl font-semibold text-white mb-4 flex items-center">
-            <span className="w-2 h-2 bg-[#A95BAB] rounded-full mr-3"></span>
-            Portfolio Images (Optional)
-          </h3>
-
-          <div className="space-y-4">
-            <label className="cursor-pointer">
-              <div className="flex flex-col items-center justify-center w-full h-32 border-2 border-gray-600/50 border-dashed rounded-lg hover:border-[#A95BAB]/50 transition-colors">
-                <Upload className="w-8 h-8 text-gray-400 mb-2" />
-                <p className="text-gray-400 text-sm">Click to upload images</p>
-                <p className="text-gray-500 text-xs">PNG, JPG, GIF up to 10MB</p>
-              </div>
-              <input
-                type="file"
-                multiple
-                accept="image/*"
-                onChange={handleImageUpload}
-                className="hidden"
-              />
-            </label>
-
-            {uploadingImages && (
-              <div className="flex items-center justify-center py-4">
-                <Loader />
-                <span className="ml-2 text-gray-300">Uploading...</span>
-              </div>
-            )}
-
-            {formData.photos.length > 0 && (
-              <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-                {formData.photos.map((photo, index) => (
-                  <div key={index} className="relative group">
-                    <img
-                      src={photo.url}
-                      alt={`Portfolio ${index + 1}`}
-                      className="w-full h-24 object-cover rounded-lg"
-                    />
-                    <button
-                      type="button"
-                      onClick={() => handleRemoveImage(index)}
-                      className="absolute top-1 right-1 bg-red-500 hover:bg-red-600 text-white rounded-full p-1 opacity-0 group-hover:opacity-100 transition-opacity"
-                    >
-                      <Trash2 className="w-3 h-3" />
-                    </button>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Actions */}
-        <div className="flex flex-col sm:flex-row gap-4 pt-6 border-t border-gray-700/50">
-          <button
-            type="button"
-            onClick={onCancel}
-            className="px-8 py-3 bg-gray-700/50 hover:bg-gray-600/50 text-gray-300 hover:text-white rounded-lg transition-all duration-300 transform hover:scale-105 font-medium"
-          >
-            Cancel
-          </button>
-          <button
-            type="submit"
-            disabled={isSubmitting || uploadingImages}
-            className={`flex-1 px-8 py-3 bg-gradient-to-r from-[#A95BAB] to-[#A95BAB]/80 text-white rounded-lg font-medium transition-all duration-300 transform hover:scale-105 ${
-              isSubmitting || uploadingImages 
-                ? "opacity-50 cursor-not-allowed" 
-                : "hover:from-[#A95BAB]/90 hover:to-[#A95BAB]/70"
-            }`}
-          >
-            {isSubmitting ? (
-              <div className="flex items-center justify-center">
-                <Loader />
-                <span className="ml-2">Creating Post...</span>
-              </div>
-            ) : (
-              "Create Post"
-            )}
-          </button>
-        </div>
-      </form>
-    </div>
-  )
-}
-
-export default CreatePostForm
