+++ frontend/src/pages/HomePage/components/ExploreSection.jsx
