+++ frontend/src/pages/auth/LoginPage/index.jsx
@@ -1,15 +1,15 @@
 "use client"
 
-import React, { useState, useEffect } from "react"
-import { Link, useNavigate, useLocation } from "react-router-dom"
+import { ArrowLeft, Eye, EyeOff } from "lucide-react"
+import { useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
-import { Eye, EyeOff, ArrowLeft } from "lucide-react"
+import { Link, useLocation, useNavigate } from "react-router-dom"
 import { Button } from "../../../components/ui/button"
+import { Card, CardContent, CardHeader, CardTitle } from "../../../components/ui/card"
 import { Input } from "../../../components/ui/input"
 import { Label } from "../../../components/ui/label"
-import { Card, CardContent, CardHeader, CardTitle } from "../../../components/ui/card"
-import { authAPI } from "../../../services/api"
-import { loginStart, loginSuccess as loginSuccessAction, loginFailure } from "../../../store/slices/authSlice"
+import { authAPI } from "../../../lib/api"
+import { loginFailure, loginStart, loginSuccess as loginSuccessAction } from "../../../store/slices/authSlice"
 
 export default function LoginPage() {
   const [email, setEmail] = useState("")
