+++ frontend/src/store/api/artistsAPI.js
